SolanaAccountManager.ts
import { Connection, PublicKey } from '@solana/web3.js';

interface Token {
    name: string;
    balance: number;
}

interface ComparisonResult {
    account1Only: Token[];
    account2Only: Token[];
    commonTokens: Token[];
}

export class SolanaAccountManager {
    private connection: Connection;

    constructor(clusterUrl: string) {
        this.connection = new Connection(clusterUrl, 'confirmed');
    }

    async fetchTokens(publicKey: string): Promise<Token[]> {
        // Mocking token retrieval for example purposes
        // Replace this with actual Solana API calls in production
        return [
            { name: "TokenA", balance: Math.floor(Math.random() * 100) },
            { name: "TokenB", balance: Math.floor(Math.random() * 100) }
        ];
    }

    async compareBalances(account1: string, account2: string): Promise<ComparisonResult> {
        const tokens1 = await this.fetchTokens(account1);
        const tokens2 = await this.fetchTokens(account2);

        const account1Only = tokens1.filter(t1 => !tokens2.some(t2 => t2.name === t1.name));
        const account2Only = tokens2.filter(t2 => !tokens1.some(t1 => t1.name === t2.name));
        const commonTokens = tokens1.filter(t1 => tokens2.some(t2 => t2.name === t1.name && t2.balance === t1.balance));

        return { account1Only, account2Only, commonTokens };
    }
}
